Let's explore an IO.pipe per runner
Each Runner starts up an IO.pipe.  The reader IO is accessible from outside
the Runner.  Running the plugin means dumping it into the writer end of
the pipe.

In rubarb itself, all the runners' reader IOs are watched with a select.
I *think* this cleanly waits for IO to take place, then does an action.
And since it's in the runner, the plugins don't care that we're dealing
with IO.  They just return strings.  Still pretty rough though, so not
positive there won't be issues.