Rendered templates ... but at what cost?
Came up with an implementation for templates.  Rubarb's DSL provides template(), which takes a block.  The block can be evaluated to interpolate a string, which includes a bunch of instance vars.  Those vars are set from the runner name/output before hand.  This implementation assumes the user will never think to use @runners or @threads...

Rails makes this work with template locals, so it should be fixable.  Not sure how.  Maybe attr_reader?