Made template work without an instance var
I think this is as clean as I'll get it.  Running the block from
template (which is just a string interpolation) in an instance_exec.
The instance exec is on an openstruct, which is just a key/value pair of
all the outputs from our plugins.  Could this also be a better way to
read in options?

Unclean options include:
* Instance vars
* Defining methods per template option
* method_missing