First take at a script runner.  Not totally satisfied.
Script runner plugin does execute a script and spit out each line as its
read.  But it relies on respawn being 0.  I think there needs to be an
explicit difference between getting the next line and restarting a
plugin that ended.  Didn't have to address that when all the ruby plugin
sytle plugins spat out a single string at a time.