Fixed script plugin
It looks simple but took a bit of tweaking to get right.

respawn() has been overridden to return immediately.  Spitting out lines
of text has been delegated to the script.  The exception is when the
script ends (as determined by IO#eof?()).  Then we run the usual respawn
sleep.

process() either returns the current process or closes a dead process and
spawns a new one.

And run() does a process.gets, not caring if the current process is
clsoed because a new one will spawn if that's the case.