Experimenting with some different config readers
Two experiments in here:

Instead of using run :plugin {options} I'm using metaprogramming to make a method per plugin name.  All it does is set up method named after each plugin to accept an array of args or a block.  (Why not both?  No idea)  The config reader returns everything it collected.  In theory it takes a string or a block, so I can use it when reading the config file as a string to collect plugin names and their blocks, and then again on each of those blocks to collect arg names.  That latter part isn't implemented yet, but *should* work.